Get countries list from carto
