+++ app/reducers/index.js
@@ -1,5 +1,5 @@
 export { default as config } from './config';
-export { default as countriesConfig } from './countriesConfig';
+export { default as countries } from './countries';
 export { default as headlines } from './headlines';
 export { default as maps } from './maps';
 export { default as modal } from './modal';
